Support 5 positions per aux channel and configuration of 12 aux channels.
Do NOT use old any version of onfgurator. Minimum configurator version is 0.51.0.
WARNING: REMOVE PROPS WHEN TESTING!
WARNING: Do not use your old aux config dumps from the CLI.

RX channels positions are:

LL < 1200
L  1200 - <1400
M  1400-1600
H  >1600 - 1800
HH >1800

5 positions, not 6 positions, were chosen because:
a) it fits in the flash better
b) it fits in ram better (all flags fit in 60 bits)
c) it’s easier to transition to, since there is still a Low Middle and High

+++ src/main/config/config_profile.h
@@ -42,7 +42,7 @@ typedef struct profile_s {
 
     uint8_t acc_unarmedcal;                 // turn automatic acc compensation on/off
 
-    uint32_t activate[CHECKBOX_ITEM_COUNT]; // activate switches, bitmask, 3 bits per channel, lower 16 bits aux1-4, upper 16 bits aux 5-8
+    uint64_t activate[CHECKBOX_ITEM_COUNT]; // activate switches, bitmask, 5 bits per channel, see rc_controls_unittest
 
     // Radio/ESC-related configuration
     uint8_t deadband;                       // introduce a deadband around the stick center for pitch and roll axis. Must be greater than zero.
